Fix RankingMetrics template parameter type. 5b1aa73e96816ac8372531a8162d9d0eac980066

+++ catboost/spark/catboost4j-spark/core/src/test/scala/ai/catboost/spark/TestMetrics.scala
@@ -27,7 +27,7 @@ object TestMetrics {
         }
       }
 
-      val actualMAPtopK = new RankingMetrics[Double](forRankingMetrics).meanAveragePrecision
+      val actualMAPtopK = new RankingMetrics[Int](forRankingMetrics).meanAveragePrecision
 
       Assert.assertEquals(expectedMAPtopK, actualMAPtopK, 1e-5)
   }
